Process arguments containing white space correctly
Fix a bug where an argument containing white space is processed as two
different arguments even if the argument was quoted correctly. Example:

$ painless-le "target dir" example.org www.example.org
> $1="target", $2="dir", $3="example.org" [...]

To ensure that white space within arguments will be processed correctly,
we must call "set" within "eval", and because of "eval" we need quoted
arguments, so remove the "-u" („unquoted“) option from getopt.